feat: add support for modules to use package.json instead of module.json
The module.json file contains many duplicative entries that are also
found in package.json.  This commit makes it possible to simply use
a package.json file (with additional BITS specific properties) as
an alternative to module.json.

This commit does not remove support for module.json files and all
existing modules will continue to works as before.

closes #6